Update message in Main.java
Change the second print statement in the Main class to output a new message. This modification was necessary to improve the readability and relevance of the output, ensuring users receive useful information instead of garbled text.

Update message in Main.java

Change the second print statement in the Main class to output a new message. This modification was necessary to improve the readability and relevance of the output, ensuring users rceive useful information instead of garbled text.